organized the files in folders
